requirements.txt
pdfplumber>=0.10.0
reportlab>=4.0.0
pyyaml>=6.0.0
